// server/src/modules/demo/demo.enums.ts
export enum OS {
	LINUX = 'Linux',
	WINDOWS = 'Windows',
	ANDROID = 'Android',
	MACOS = 'Mac OS',
	IOS = 'IOS'
}

export enum Browser {
	CHROME = 'Chrome',
	FIREFOX = 'Firefox',
	SAFARI = 'Safari',
	OPERA = 'Opera',
	EDGE = 'Microsoft Edge'
}

export enum Platform {
	DESKTOP = 'desktop',
	MOBILE = 'mobile'
}